* src/devices/grolbp/lbp.h, src/preproc/grn/gprint.h, src/preproc/grn/hgraph.cpp, src/preproc/grn/hpoint.cpp, src/preproc/html/pushback.cpp, src/preproc/html/pushback.h: Use `double' instead of `float' everywhere. * src/preproc/grn/main.cpp: Use `double' instead of `float' everywhere. (main): Add return value. * src/preproc/grn/hdb.cpp: Use `double' instead of `float' everywhere. Update all user functions.
* src/include/search.h, src/include/ptable.h, src/include/printer.h,
src/preproc/eqn/box.h, src/preproc/pic/object.h,
src/preproc/refer/refer.h, src/preproc/tbl/table.h,
src/preproc/tbl/table.cpp, src/roff/troff/env.h,
src/roff/troff/div.h, src/roff/troff/token.h, src/roff/troff/node.h,
src/roff/troff/input.cpp, src/roff/troff/request.h,
src/roff/troff/node.cpp: Don't mix `struct' and `class' in forward
declarations.

* src/utils/indxbib/signal.c: Include stdlib.h.

+++ src/devices/grolbp/lbp.h
@@ -348,11 +348,11 @@ static inline void
 splinerel(double px,double py,int flush)
 {
   static int lx = 0 ,ly = 0;
-  static float pend = 0.0;
+  static double pend = 0.0;
   static int dy = 0, despx = 0, despy = 0, sigpend = 0;
   int dxnew ,dynew, sg;
   char xcoord[4],ycoord[4];
-  float npend ;
+  double npend ;
 
   if (flush == -1) {lx = (int)px; ly = (int)py; return;};
